requirements.txt
fastapi
uvicorn
python-dotenv
odmantic
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
